LPD-55683 Move Helper class from util to helper package
